fix UUID FileBox bug
