+++ package.json
@@ -55,9 +55,11 @@
     "@types/google-protobuf": "^3.15.5",
     "@types/lru-cache": "^7.5.0",
     "@types/semver": "^7.3.9",
+    "@types/temp": "^0.9.1",
     "@types/uuid": "^8.3.4",
     "@types/ws": "^8.5.3",
     "get-port": "^6.1.2",
+    "temp": "^0.9.4",
     "typed-emitter": "^1.5.0-from-event",
     "utility-types": "^3.10.0",
     "wechaty-puppet-mock": "^1.18.2",
@@ -67,8 +69,9 @@
     "wechaty-puppet": "^1.19.1"
   },
   "dependencies": {
-    "clone-class": "^1.0.3",
+    "clone-class": "^1.1.1",
     "ducks": "^1.0.2",
+    "file-box": "^1.5.5",
     "flash-store": "^1.3.4",
     "gerror": "^1.0.16",
     "redux-observable": "^2.0.0",
